+++ spectrochempy/examples/analysis/a_decomposition/plot_efa_keller_massart.py
@@ -66,10 +66,10 @@ efa.fit(dataset)
 # %%
 # Plots of the log(EV) for the forward and backward analysis
 #
-efa.f_ev.T.plot(yscale="log", legend=efa.f_ev.x.labels)
+efa.f_ev.T.plot(yscale="log", legend=efa.f_ev.k.labels)
 
 # %%
-efa.b_ev.T.plot(yscale="log", legend=efa.b_ev.x.labels)
+efa.b_ev.T.plot(yscale="log", legend=efa.b_ev.k.labels)
 
 # %%
 # Looking at these EFA curves, it is quite obvious that only two components