+++ tests/test_core/test_readers/test_read_topspin.py
@@ -26,21 +26,21 @@ NMRDATA = DATADIR / "nmrdata"
 def test_readtopspin():
     # A.plot()
     nd = scp.read_topspin(nmrdir / "exam2d_HC/3/pdata/1/2rr")
-    assert str(nd) == "NDDataset: [quaternion] unitless (shape: (y:1024, x:1024))"
+    assert str(nd) == "NDDataset: [quaternion] pp (shape: (y:1024, x:1024))"
 
     # Select a TOPSPIN spectra using the full name
     nd = scp.read_topspin(nmrdir / "topspin_1d/1/fid")
-    assert str(nd) == "NDDataset: [complex128] unitless (size: 12411)"
+    assert str(nd) == "NDDataset: [complex128] pp (size: 12411)"
 
     nd = scp.read_topspin(nmrdir / "topspin_1d/1/pdata/1/1r")
-    assert str(nd) == "NDDataset: [complex128] unitless (size: 16384)"
+    assert str(nd) == "NDDataset: [complex128] pp (size: 16384)"
 
     # Select a TOPSPIN spectra using the full name
     nd = scp.read_topspin(nmrdir / "topspin_2d/1/ser")
-    assert str(nd) == "NDDataset: [quaternion] unitless (shape: (y:96, x:948))"
+    assert str(nd) == "NDDataset: [quaternion] pp (shape: (y:96, x:948))"
 
     nd = scp.read_topspin(nmrdir / "topspin_2d/1/pdata/1/2rr")
-    assert str(nd) == "NDDataset: [quaternion] unitless (shape: (y:1024, x:2048))"
+    assert str(nd) == "NDDataset: [quaternion] pp (shape: (y:1024, x:2048))"
 
     # alternative syntax
     nd1 = scp.read_topspin(nmrdir / "topspin_2d", expno=1, procno=1)
@@ -52,8 +52,8 @@ def test_readtopspin():
 
     nd = scp.read_topspin(nmrdir, glob="topspin*/*/pdata/*/*")
     assert isinstance(nd, list)
-    assert str(nd[0]) == "NDDataset: [complex128] unitless (size: 16384)"
-    assert str(nd[1]) == "NDDataset: [quaternion] unitless (shape: (y:1024, x:2048))"
+    assert str(nd[0]) == "NDDataset: [complex128] pp (size: 16384)"
+    assert str(nd[1]) == "NDDataset: [quaternion] pp (shape: (y:1024, x:2048))"
 
 
 # @pytest.mark.skipif(