+++ docs/whatsnew/changelog.rst
@@ -22,7 +22,13 @@ New features
 
 * Fancy indexing using location now supported.
 * Add an example for NMR processing of a series of CP-MAS spectra.
+* Add an example for processing NMR relaxation data
+* Add an option to `read_topspin` to create ``y`` coordinates from a ``vdlist`` file.
 * Add option to plot to add markers on curves
+* Add a new method to the `Optimize` class to perform a least-square fitting. It is
+  based on the `scipy.optimize.least_squares` function, allowing much faster operation
+  for simple curve fitting
+* Add the possibility to define user-defined functions in the `Optimize` class.
 
 .. section
 
@@ -32,6 +38,8 @@ Bug fixes
 
 * Sorting coordinates now work with multi-coordinates axis.
 * Fix a bug when concatenating datasets with multi-coordinates axis.
+* Fix a bug in coordset definition for integration methods.
+* Fix coordinates definitions in Analysis methods.
 
 .. section